+++ crates/objc2-encode/src/lib.rs
@@ -48,7 +48,7 @@
 #![warn(clippy::missing_errors_doc)]
 #![warn(clippy::missing_panics_doc)]
 // Update in Cargo.toml as well.
-#![doc(html_root_url = "https://docs.rs/objc2-encode/3.0.0")]
+#![doc(html_root_url = "https://docs.rs/objc2-encode/4.0.0")]
 #![cfg_attr(feature = "unstable-c-unwind", feature(c_unwind))]
 
 #[cfg(doctest)]